requirements.txt
streamlit
google-genai
wave
pyaudio